Phase 6 Step 8 (part of #757): collapse dual-engine shadow-comparison shell (v0.6.74)
Every subsystem's legacy/FSM cutover flag has been removed (Steps 1-7), so
the coordinator's two live AutomationEngine instances (_engine_a/_engine_b)
had become behaviorally identical. This collapses them to a single engine:

- Removes the dual construction, _shadow_is_primary routing, the
  automation_engine/shadow_automation_engine properties, _sync_shadow_inputs(),
  _apply_engine_roles(), async_set_shadow_engine_primary(), _mirror_to_shadow(),
  _update_shadow_engine_diagnostic(), and the shadow_engine_diagnostic property.
- Removes switch.py's ClimateAdvisorShadowEnginePrimarySwitch and sensor.py's
  ClimateAdvisorShadowEngineStatusSensor.
- Removes tools/sim_harness/shadow_engine_pair.py and its isolation-proof tests.
- 8 call sites that fed the override/grace FSM as a side effect of
  _mirror_to_shadow()'s finally block were rewired to feed it directly via a
  new _feed_override_grace_fsm_on_detect() helper, preserving that still-live
  behavior exactly (verified line-by-line against the removed function).
- Removed tools/build_historical_scenario.py's --type shadow_disagreement mode
  (dead code once the diagnostic that fed it no longer exists).
- Relocated two still-relevant tests that lived in the otherwise shadow-only
  test_shadow_engine_coverage.py into test_override_grace_fsm.py and
  test_fan_fsm.py before deleting that file.

Executed via an Investigator -> Executor -> Verifier pipeline; the Verifier
caught the dead build_historical_scenario.py tool mode and sent it back
before this was considered done. Full suite: 4822 passed, 6 skipped. Golden
corpus: 91/91. ruff clean.

Remaining in Phase 6: Step 9 (final verification, soak, close-out).
