+++ custom_components/climate_advisor/api.py
@@ -632,10 +632,11 @@ class ClimateAdvisorResumeFromPauseView(HomeAssistantView):
             return self.json({"status": "ok", "message": "Not currently paused"})
 
         restored_mode = await ae.resume_from_pause()
-        # Issue #615: mirror onto the shadow engine too — resume_from_pause() internally
-        # no-ops if the shadow doesn't believe it's paused, same unconditional-mirror
-        # pattern used for the coordinator-driven decision methods.
-        await coordinator._mirror_to_shadow("resume_from_pause")
+        # Issue #757 Phase 6 Step 8: dual-engine shell removed; this call used to also
+        # mirror onto the shadow engine, whose _mirror_to_shadow() finally block secretly
+        # fed the override/grace FSM's DASHBOARD_RESUME entry event as a side effect.
+        # Feed it directly now that the shadow engine is gone.
+        coordinator._feed_override_grace_fsm_on_detect("resume_from_pause")
         return self.json(
             {
                 "status": "ok",
